+++ src/main/java/Covid19project/Repository/AppointmentRepository/IAppointmentRepo.java
@@ -14,7 +14,7 @@ public interface IAppointmentRepo {
 
     Appointment findAppointmentById(int apptID);
 
-    Appointment findAppointmentByCpr(Integer cpr);
+    Appointment findAppointmentByCpr(Long cpr);
 
     //Update
     Appointment updateAppointment(int apptID, Appointment appointment);
